+++ src/main/java/com/store/api/dto/ProductResponse.java
@@ -1,7 +1,6 @@
 package com.store.api.dto;
 
 import com.store.api.domain.Product;
-import com.store.api.domain.ProductOrders;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -12,7 +11,7 @@ import java.util.UUID;
 public class ProductResponse {
     private UUID productCode;
     private String name;
-    private Integer productQuantity;
+    private Long productQuantity;
     private BigDecimal price;
 
     public static ProductResponse converter(Product product) {
@@ -24,5 +23,4 @@ public class ProductResponse {
 
         return productResponse;
     }
-
 }
